requirements.txt
toml
pandas
numpy
moviepy
PyQt5
pydub
portion
openai-whisper
opencc
multipledispatch
opencv-python
srt